+++ src/library/renderhud/RenderHUD_SDL1.cpp
@@ -33,7 +33,7 @@ namespace orig {
             Uint32 Rmask, Uint32 Gmask, Uint32 Bmask, Uint32 Amask);
     static void (*SDL_FreeSurface)(SDL1::SDL_Surface *surface);
     static SDL1::SDL_Surface *(*SDL_GetVideoSurface)(void);
-    static int (*SDL_BlitSurface)(SDL1::SDL_Surface *src, SDL1::SDL_Rect *srcrect, SDL1::SDL_Surface *dst, SDL1::SDL_Rect *dstrect);
+    static int (*SDL_UpperBlit)(SDL1::SDL_Surface *src, SDL1::SDL_Rect *srcrect, SDL1::SDL_Surface *dst, SDL1::SDL_Rect *dstrect);
     static uint8_t (*SDL_SetClipRect)(SDL1::SDL_Surface *surface, const SDL1::SDL_Rect *rect);
     static void (*SDL_GetClipRect)(SDL1::SDL_Surface *surface, SDL1::SDL_Rect *rect);
 }
@@ -47,7 +47,7 @@ void RenderHUD_SDL1::renderText(const char* text, Color fg_color, Color bg_color
     LINK_NAMESPACE_SDL1(SDL_CreateRGBSurfaceFrom);
     LINK_NAMESPACE_SDL1(SDL_FreeSurface);
     LINK_NAMESPACE_SDL1(SDL_GetVideoSurface);
-    LINK_NAMESPACE_SDL1(SDL_BlitSurface);
+    LINK_NAMESPACE_SDL1(SDL_UpperBlit);
     LINK_NAMESPACE_SDL1(SDL_SetClipRect);
     LINK_NAMESPACE_SDL1(SDL_GetClipRect);
 
@@ -66,7 +66,7 @@ void RenderHUD_SDL1::renderText(const char* text, Color fg_color, Color bg_color
     SDL1::SDL_Rect clip_rect;
     orig::SDL_GetClipRect(screen, &clip_rect);
     orig::SDL_SetClipRect(screen, nullptr);
-    orig::SDL_BlitSurface(sdlsurf, NULL, screen, &rect);
+    orig::SDL_UpperBlit(sdlsurf, NULL, screen, &rect);
     orig::SDL_SetClipRect(screen, &clip_rect);
 
     orig::SDL_FreeSurface(sdlsurf);